+++ src/main/java/org/cook/booking_system/service/implementation/images/RoomTypeImageServiceImpl.java
@@ -2,12 +2,13 @@ package org.cook.booking_system.service.implementation.images;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.cook.booking_system.entity.RoomTypeEntity;
 import org.cook.booking_system.entity.images.RoomTypeImageEntity;
 import org.cook.booking_system.mapper.images.RoomTypeImageMapper;
-import org.cook.booking_system.model.images.RoomTypeImage;
+import org.cook.booking_system.model.images.Image;
 import org.cook.booking_system.repository.RoomTypeRepository;
 import org.cook.booking_system.repository.images.RoomTypeImageRepository;
-import org.cook.booking_system.service.service_interface.images.RoomTypeImageService;
+import org.cook.booking_system.service.service_interface.images.ImageService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -15,34 +16,43 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class RoomTypeImageServiceImpl implements RoomTypeImageService {
+public class RoomTypeImageServiceImpl implements ImageService {
 
     private final RoomTypeImageRepository roomTypeImageRepository;
     private final RoomTypeRepository roomTypeRepository;
     private final RoomTypeImageMapper roomTypeImageMapper;
 
     @Transactional
-    public RoomTypeImage addImage(Long roomTypeId, String imageUrl) {
-        var roomType = roomTypeRepository.findById(roomTypeId)
-                .orElseThrow(() -> new EntityNotFoundException("Room type not found with id -> " + roomTypeId));
+    public Image addImage(Long roomTypeId, String imageUrl) {
+        RoomTypeEntity roomType = roomTypeRepository.findByIdAndDeletedFalse(roomTypeId)
+                .orElseThrow(() -> new EntityNotFoundException("Room type not found or marked as deleted with id -> " + roomTypeId));
 
-        RoomTypeImageEntity entity = new RoomTypeImageEntity();
-        entity.setRoomType(roomType);
-        entity.setUrl(imageUrl);
+        RoomTypeImageEntity image = new RoomTypeImageEntity();
+        image.setRoomType(roomType);
+        image.setUrl(imageUrl);
 
-        return roomTypeImageMapper.toModel(roomTypeImageRepository.save(entity));
+        return roomTypeImageMapper.toModel(roomTypeImageRepository.save(image));
     }
 
     @Transactional(readOnly = true)
-    public List<RoomTypeImage> getImages(Long roomTypeId) {
-        return roomTypeImageRepository.findByRoomTypeId(roomTypeId)
-                .stream()
+    public List<Image> getImages(Long roomTypeId) {
+        RoomTypeEntity roomType = roomTypeRepository.findById(roomTypeId)
+                .orElseThrow(() -> new EntityNotFoundException("Room type not found with id -> " + roomTypeId));
+
+        return roomTypeImageRepository.findByRoomTypeId(roomTypeId).stream()
                 .map(roomTypeImageMapper::toModel)
                 .toList();
     }
 
     @Transactional
-    public void deleteImage(Long id) {
+    public void deleteImage(Long roomTypeId, Long id) {
+        RoomTypeEntity roomType = roomTypeRepository.findByIdAndDeletedFalse(roomTypeId)
+                .orElseThrow(() -> new EntityNotFoundException("Room type not found or marked as deleted with id -> " + roomTypeId));
+
+        if (!roomTypeImageRepository.existsByIdAndRoomTypeId(id, roomTypeId)) {
+            throw new EntityNotFoundException("Image not found with id -> " + id);
+        }
+
         roomTypeImageRepository.deleteById(id);
     }
 }
